refactored virtual service related functions from route to virtual service
